+++ apps/gateway/src/modules/logistic/crafts/travels/travels.controller.ts
@@ -13,10 +13,16 @@ import {
   UseInterceptors,
   UsePipes,
 } from '@nestjs/common';
-import { TravelDataSerializer, TravelItemsSerializer, TravelSerializer } from '@app/common/serializers/logistic';
+import {
+  RoutingDataSerializer,
+  RoutingSerializer,
+  TravelDataSerializer,
+  TravelItemsSerializer,
+  TravelSerializer,
+} from '@app/common/serializers/logistic';
+import { RoutingDto, CreateTravelDto, CreateTravelItemsDto, UpdateTravelDto } from '@app/common/dto/logistic';
 import { GatewayInterceptors, ResponseInterceptors, WriteInterceptors } from '@app/common/core/interceptors';
 import { Audit, Cache, RateLimit, SetPolicy, SetScope, Validation } from '@app/common/core/metadatas';
-import { CreateTravelDto, CreateTravelItemsDto, UpdateTravelDto } from '@app/common/dto/logistic';
 import { AuthorityInterceptor, ProjectionInterceptor } from '@app/common/core/interceptors/mongo';
 import { ApiBearerAuth, ApiParam, ApiQuery, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { FilterDto, FilterOneDto, QueryFilterDto } from '@app/common/core/dto/mongo';
@@ -24,6 +30,7 @@ import { Controller as ControllerClass } from '@app/common/core/classes/mongo';
 import { Controller as IController } from '@app/common/core/interfaces/mongo';
 import { AuthGuard, PolicyGuard, ScopeGuard } from '@app/common/core/guards';
 import { Action, COLLECTION, Resource, Scope } from '@app/common/core';
+import { getSseMessage, mapToInstance } from '@app/common/core/utils';
 import { Travel, TravelDto } from '@app/common/interfaces/logistic';
 import { LogisticProvider } from '@app/common/providers/logistic';
 import { Filter, Meta, Perm } from '@app/common/core/decorators';
@@ -31,9 +38,8 @@ import { AllExceptionsFilter } from '@app/common/core/filters';
 import { TotalSerializer } from '@app/common/core/serializers';
 import { SentryInterceptor } from '@ntegral/nestjs-sentry';
 import { ValidationPipe } from '@app/common/core/pipes';
-import { getSseMessage } from '@app/common/core/utils';
 import { Metadata } from '@app/common/core/interfaces';
-import { Observable, switchMap } from 'rxjs';
+import { from, Observable, switchMap } from 'rxjs';
 import { Permission } from 'abacl';
 import { Response } from 'express';
 
@@ -51,6 +57,20 @@ export class TravelsController extends ControllerClass<Travel, TravelDto> implem
   constructor(readonly provider: LogisticProvider) {
     super(provider.travels, TravelSerializer);
   }
+
+  @Post('routing')
+  @Audit('GATEWAY')
+  @Cache(COLL_PATH, 'fill')
+  @SetScope(Scope.ResolveLogisticTravels)
+  @ApiResponse({ type: RoutingDataSerializer })
+  @SetPolicy(Action.Resolve, Resource.LogisticTravels)
+  routing(@Meta() meta: Metadata, @Body() data: RoutingDto): Observable<RoutingDataSerializer> {
+    return from(this.provider.travels.routing(data, { meta })).pipe(mapToInstance(RoutingSerializer, 'data'));
+  }
+
+  // ##############################
+  // Common Method's
+  // ##############################
 
   @Get('count')
   @Cache(COLL_PATH, 'fill')